refactor: 최상단에 도메인이 위치하도록 폴더 구조 Refactoring
