--- mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/adapter/converter/in/AuthCommandConverter.java
@@ -1,13 +1,0 @@
-package org.rookedsysc.mybatishexagonalexam.adapter.converter.in;
-
-import org.rookedsysc.mybatishexagonalexam.adapter.in.AuthRequest;
-import org.rookedsysc.mybatishexagonalexam.application.port.in.AuthCommand;
-
-public class AuthCommandConverter {
-    public static AuthCommand toCommand(AuthRequest request) {
-        return AuthCommand.builder()
-            .email(request.email())
-            .password(request.password())
-            .build();
-    }
-}
