--- mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/adapter/in/AuthRequest.java
@@ -1,7 +1,0 @@
-package org.rookedsysc.mybatishexagonalexam.adapter.in;
-
-public record AuthRequest(
-    String email,
-    String password
-) {
-}
