--- mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/application/port/in/AuthCommand.java
@@ -1,10 +1,0 @@
-package org.rookedsysc.mybatishexagonalexam.application.port.in;
-
-import lombok.Builder;
-
-@Builder
-public record AuthCommand(
-    String email,
-    String password
-) {
-}
