--- mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/application/port/in/PostCommand.java
@@ -1,8 +1,0 @@
-package org.rookedsysc.mybatishexagonalexam.application.port.in;
-
-public record PostCommand(
-    Long userId,
-    String title,
-    String content
-) {
-}
