--- mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/application/port/in/PostUploadUsecase.java
@@ -1,10 +1,0 @@
-package org.rookedsysc.mybatishexagonalexam.application.port.in;
-
-import org.rookedsysc.mybatishexagonalexam.adapter.out.entity.PostEntity;
-
-public interface PostUploadUsecase {
-    /**
-     * 업로드 한 포스트의 ID를 반환
-     */
-    long upload(PostCommand command);
-}
