--- mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/application/port/in/SignUpUseCase.java
@@ -1,7 +1,0 @@
-package org.rookedsysc.mybatishexagonalexam.application.port.in;
-
-import org.rookedsysc.mybatishexagonalexam.adapter.out.entity.UserEntity;
-
-public interface SignUpUseCase {
-    UserEntity signUp(AuthCommand command);
-}
