+++ mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/auth/adapter/in/AuthRequest.java
@@ -1,0 +1,7 @@
+package org.rookedsysc.mybatishexagonalexam.auth.adapter.in;
+
+public record AuthRequest(
+    String email,
+    String password
+) {
+}
