+++ mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/auth/application/port/in/AuthCommand.java
@@ -1,0 +1,10 @@
+package org.rookedsysc.mybatishexagonalexam.auth.application.port.in;
+
+import lombok.Builder;
+
+@Builder
+public record AuthCommand(
+    String email,
+    String password
+) {
+}
