+++ mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/auth/application/port/in/SignUpUseCase.java
@@ -1,0 +1,7 @@
+package org.rookedsysc.mybatishexagonalexam.auth.application.port.in;
+
+import org.rookedsysc.mybatishexagonalexam.auth.adapter.out.entity.UserEntity;
+
+public interface SignUpUseCase {
+    UserEntity signUp(AuthCommand command);
+}
