+++ mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/post/application/port/in/PostCommand.java
@@ -1,0 +1,8 @@
+package org.rookedsysc.mybatishexagonalexam.post.application.port.in;
+
+public record PostCommand(
+    Long userId,
+    String title,
+    String content
+) {
+}
