+++ mybatis-hexagonal-exam/src/main/java/org/rookedsysc/mybatishexagonalexam/post/application/port/in/PostUploadUsecase.java
@@ -1,0 +1,8 @@
+package org.rookedsysc.mybatishexagonalexam.post.application.port.in;
+
+public interface PostUploadUsecase {
+    /**
+     * 업로드 한 포스트의 ID를 반환
+     */
+    long upload(PostCommand command);
+}
