+++ mybatis-hexagonal-exam/src/main/resources/mapper/PostMapper.xml
@@ -2,10 +2,10 @@
 <!DOCTYPE mapper
     PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
     "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.rookedsysc.mybatishexagonalexam.adapter.out.PostMapper">
+<mapper namespace="org.rookedsysc.mybatishexagonalexam.post.adapter.out.PostMapper">
 
     <!-- Insert Post -->
-    <insert id="save" parameterType="org.rookedsysc.mybatishexagonalexam.adapter.out.entity.PostEntity" useGeneratedKeys="true" keyProperty="id">
+    <insert id="save" parameterType="org.rookedsysc.mybatishexagonalexam.post.adapter.out.entity.PostEntity" useGeneratedKeys="true" keyProperty="id">
         INSERT INTO post (user_id, title, content)
         VALUES (#{userId}, #{title}, #{content})
     </insert>
